\name{randomiseTimestamp}
\alias{randomiseTimestamp}
\title{create a random order for the timestamps}
\usage{
  randomiseTimestamp(timestamp, min.difference = 10)
}
\arguments{
  \item{timestamp}{a vector of timestamps}

  \item{min.difference}{the minimal difference between two
  timestamps in seconds.}
}
\value{
  a vector with the new order of the timestamps
}
\description{
  create a random order for the timestamps
}

